grammar implementation

+++ sources/com.nextbreakpoint.nextfractal.flux/src/main/java/com/nextbreakpoint/nextfractal/flux/PaletteExpression.java
@@ -2,5 +2,5 @@ package com.nextbreakpoint.nextfractal.flux;
 
 @FunctionalInterface
 public interface PaletteExpression {
-	public double evaluate(int start, int end, int step);
+	public double evaluate(double start, double end, double step);
 }
